requirements.txt
Django==5.1.7
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-writable-nested==0.7.0
environ==1.0
httpcore==1.0.7
httpx==0.25.2
pillow==10.4.0
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.22.0
PyJWT==2.9.0
python-telegram-bot==20.6
requests==2.32.3
sqlparse==0.5.1
fastapi==0.115.12

